Move around the test configuration files into the src/test/resources directory. Also, ignore the tests when running unit tests
