--- service/application/src/test/java/com/solace/maas/ep/event/management/agent/commandManager/badClasses/CommandTypeBad.java
@@ -1,7 +1,0 @@
-package com.solace.maas.ep.event.management.agent.commandManager.badClasses;
-
-public enum CommandTypeBad {
-    terraform,
-    badEnum
-}
-
