+++ service/application/src/test/java/com/solace/maas/ep/event/management/agent/realTests/badClasses/CommandTypeBad.java
@@ -1,0 +1,7 @@
+package com.solace.maas.ep.event.management.agent.realTests.badClasses;
+
+public enum CommandTypeBad {
+    terraform,
+    badEnum
+}
+
